Refactor all api calls to it's own module, add logging, fix logging levels
